+++ protocol/response.game.h
@@ -68,8 +68,8 @@ public:
             auto user = pair.second->data();
             auto& name = user->name();
 
-            out_stream.write_u8(0x10 * user->nation())
-                      .write_u8(0x10 * user->promotion())
+            out_stream.write_u8(0x10 * static_cast<int>(user->nation()))
+                      .write_u8(0x10 * static_cast<int>(user->promotion()))
                       .write_u8((&this->me == user) ? 0x88 : 0x0F)
                       .write(name, false);
         }
